+++ CHANGES.md
@@ -21,6 +21,7 @@ v2.4.20 - YYYY-MM-DD
 - Added validation of IPP "printer-state-reasons" and
   "printer-mandatory-job-attributes" attributes (Issue #1632)
 - Added missing Set-Printer-Attributes policy to cupsd.conf.
+- Removed problematic debug printfs from `dnssd` backend.
 - Fixed mapping of standard PPD/PWG/legacy media size names to the local PPD
   size name (Issue #1375)
 - Fixed handling of multiple PPD: keywords from filters (Issue #1562,